rename: Enum타입 Entity 패키지로 이동
