+++ src/main/java/gladiator/philosopher/comment/entity/Comment.java
@@ -1,7 +1,6 @@
 package gladiator.philosopher.comment.entity;
 
 import gladiator.philosopher.account.entity.Account;
-import gladiator.philosopher.comment.dto.CommentStatus;
 import gladiator.philosopher.common.entity.BaseEntity;
 import gladiator.philosopher.mention.entity.Mention;
 import gladiator.philosopher.recommend.entity.Recommend;
