+++ src/main/java/gladiator/philosopher/thread/repository/ThreadCustomRepositoryImpl.java
@@ -9,11 +9,11 @@ import com.querydsl.core.types.dsl.Wildcard;
 import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import gladiator.philosopher.recommend.entity.QRecommend;
-import gladiator.philosopher.thread.dto.Sort;
 import gladiator.philosopher.thread.dto.ThreadResponseDto;
 import gladiator.philosopher.thread.dto.ThreadSearchCond;
 import gladiator.philosopher.thread.dto.ThreadSimpleResponseDto;
 import gladiator.philosopher.thread.entity.QThread;
+import gladiator.philosopher.thread.entity.Sort;
 import gladiator.philosopher.thread.entity.Thread;
 import gladiator.philosopher.thread.entity.ThreadLocation;
 import java.util.List;
